Require Jenkins 2.387.3 or newer
Require Jenkins 2.387.3 or newer so that users do not think that versions
older than 2.387.3 are actively tested.

https://www.jenkins.io/doc/developer/plugin-development/choosing-jenkins-baseline/

Choosing a Jenkins version to build against recommends 2.387.3 or 2.375.4
as the base version.  Versions prior to 2.375.4 have known security
vulnerabilities in the plugin manager.  I no longer test anything older
than 2.387.3.

As of 1 May 2023, 99% of installs of the most recent release
(448. released 6 weeks ago) are running 2.387.1 or newer.  Over 70%
of the installs of the release from Feb 2023 are running 2.387.1 or newer.