Use DateTimeFormatter if DateFormat does not parse
Java 21 changed the behavior of the DateFormat parsing to no longer
accept strings that contain only a time.  Previous versions of Java
would default to January 1, 1970 if no date value was specified.

The Java 21 behavior is quite reasonable, since a Date specifies a number
of milliseconds since the epoch, not an offset since the start of the day.
The plugin misuses a Date object to represent what Java 8 declared as
LocalTime, a representation of time within a day.

This change continues with the previous behavior, checking a DateFormat
first.  If the DateFormat parse fails, then it iterates through a series
of DateTimeFormatter patterns to check if the passed string matches one
of those formats.

I've not detected any case in the typical use of the plugin where the
new Java 21 behavior affects plugin behavior.  It seems that there should
be cases where the change is relevant, but I was not able to detect any
of those cases.