[JENKINS-72653] fix timezone handling
the implementation to prefill the scheduled time with the global value
was not working properly, when the timezone of the controller was
different from the timezone that was configured. The implementation used
java.util.Date which has only limited capabilites to work with
timeszones.
This change replaces all date/time handling with new classes from the
java.time packages.
The timezone is no longer a free text field but a dropdown that gets
filled with all available zone ids.
The datetime field to schedule a build is no longer locale dependent but
uses a fixed format based on 24h format (see [JENKINS-72645]).
It is optional to enter seconds.
The actual timezone that will be used is shown as field description.
The time input on global config allows to specify am/pm but also 24h
formats, seconds are optional.