Remove developers section from pom file (#379)
The repository permissions updater is the definitive list of plugin
developers.  The plugin archetype has not suggested a `<developers>`
section since the 2022 merge of:

* https://github.com/jenkinsci/archetypes/pull/515

Rather than carry duplicated information in the pom.xml file, remove the
`<developers>` section to better match the plugin archetype and further
reduce the maintenance burden.